refactor: RestaurnatList 내부 훅을 커스텀 훅으로 분리
